// src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type PublicEnumProofingStatus =
  | "Out for Proof"
  | "Response from Client"
  | "Approved"
  | "Disapproved"
  | "Canceled"
export type PublicEnumTaskPriority = "Low" | "Medium" | "High"
export type PublicEnumTaskStatus = "To-Do" | "In Progress" | "Done" | "On Hold"
export type PublicEnumTeamRole = "Admin" | "Member" | "Viewer"

export interface Database {
  public: {
    Tables: {
// ...
      contacts: {
        Row: {
          created_at: string
          email: string | null
          id: number
          name: string | null
          notes: string | null
          owner_id: string
          page_id: string | null
          source: string | null
          status: string | null
        }
        Insert: {
          created_at?: string
          email?: string | null
          id?: number
          name?: string | null
          notes?: string | null
          owner_id: string
          page_id?: string | null
          source?: string | null
          status?: string | null
        }
        Update: {
          created_at?: string
          email?: string | null
          id?: number
          name?: string | null
          notes?: string | null
          owner_id?: string
          page_id?: string | null
          source?: string | null
          status?: string | null
        }
        Relationships: []
      }
// ...
      proofing_comments: {
        Row: {
          author_name: string
          comment_text: string
          created_at: string
          id: number
          metadata: Json | null
          proofing_request_id: string
          user_id: string | null
        }
        Insert: {
          author_name: string
          comment_text: string
          created_at?: string
          id?: number
          metadata?: Json | null
          proofing_request_id: string
          user_id?: string | null
        }
        Update: {
          author_name?: string
          comment_text?: string
          created_at?: string
          id?: number
          metadata?: Json | null
          proofing_request_id?: string
          user_id?: string | null
        }
        Relationships: []
      }
      proofing_requests: {
        Row: {
          client_id: number | null
          created_at: string
          description: string | null
          id: string
          related_entity_id: string | null
          related_entity_type: string | null
          status: PublicEnumProofingStatus
          title: string
          user_id: string
          versions: Json | null
        }
        Insert: {
          client_id?: number | null
          created_at?: string
          description?: string | null
          id?: string
          related_entity_id?: string | null
          related_entity_type?: string | null
          status?: PublicEnumProofingStatus
          title: string
          user_id: string
          versions?: Json | null
        }
        Update: {
          client_id?: number | null
          created_at?: string
          description?: string | null
          id?: string
          related_entity_id?: string | null
          related_entity_type?: string | null
          status?: PublicEnumProofingStatus
          title?: string
          user_id?: string
          versions?: Json | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      proofing_status: PublicEnumProofingStatus
      task_priority: PublicEnumTaskPriority
      task_status: PublicEnumTaskStatus
      team_role: PublicEnumTeamRole
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Row"]
export type TablesInsert<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Insert"]
export type TablesUpdate<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Update"]